Added Boss Room Initialization
Spawns in most recently generated room after X seconds, normally this should be the furthest room from the center. If an extra long generation is made, it's possible the boss room will not be the last room in the path, but likely will be implemented to end that level once boss is defeated, so it shouldn't matter.